requirements.txt
PyPDF2==1.26.0
python-docx==0.8.11
werkzeug==2.2.3
flask-uploads==0.2.1
flask-mail==0.9.1
